requirements.txt
amqp==5.0.5
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
configparser==5.0.2
contextlib2==0.6.0.post1
dcdatabase==1.2
enum34==1.1.10
funcsigs==1.0.2
idna==2.10
importlib-metadata==3.7.2
kombu==5.0.2
lxml==4.6.2
mccabe==0.6.1
more-itertools==8.7.0
numpy==1.18.5
pathlib2==2.3.5
prompt-toolkit==3.0.16
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scandir==1.10.0
scipy==1.4.1
selenium==3.141.0
sentinels==1.0.0
six==1.15.0
soupsieve==2.2
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1